Rename all mark/stat modules to be singular (#2941)
