+++ seaborn/_stats/histogram.py
