+++ seaborn/objects.py
@@ -30,14 +30,14 @@ from seaborn._core.plot import Plot  # noqa: F401
 
 from seaborn._marks.base import Mark  # noqa: F401
 from seaborn._marks.area import Area, Ribbon  # noqa: F401
-from seaborn._marks.bars import Bar, Bars  # noqa: F401
-from seaborn._marks.lines import Line, Lines, Path, Paths, Interval  # noqa: F401
+from seaborn._marks.bar import Bar, Bars  # noqa: F401
+from seaborn._marks.line import Line, Lines, Path, Paths, Interval  # noqa: F401
 from seaborn._marks.scatter import Dot, Scatter  # noqa: F401
 
 from seaborn._stats.base import Stat  # noqa: F401
 from seaborn._stats.aggregation import Agg, Est  # noqa: F401
 from seaborn._stats.regression import OLSFit, PolyFit  # noqa: F401
-from seaborn._stats.histograms import Hist  # noqa: F401
+from seaborn._stats.histogram import Hist  # noqa: F401
 
 from seaborn._core.moves import Dodge, Jitter, Norm, Shift, Stack  # noqa: F401
